Enumerate Node.js APIs on main Node.js docs page
We don't currently have one place where you can look at-a-glance at what Node.js APIs are supported in Workers.

The compatibility matrix here is very cool, overly detailed for a quick reference on what is or isn't supported:

https://workers-nodejs-compat-matrix.pages.dev/

The clearest thing we can do is to provide a simple list of all non-deprecated, non-experimental APIs from Node.js (pulled this straight from the Node.js docs, in the same alphabetical order) — and state clearly which are supported, which we have partial support for, and which are not supported. And then link to our own relevant docs page.

There are a few here that are currently supported, but we haven't yet added a docs page for. We'll need to follow up on this, so that the sidebar includes all of the Node.js APIs we currently support in the Workers Runtime.